chore(kms): rename file
